realtime_ai_backend/backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
asyncio
